Add full tag functionality and visual waveform functionality
This is a huge push which completely integrates all of the tagging system features desired except for dragging to move them, as well as implementing the waveform visualization, addition of markers wherever the playhead currently is, and double clicking on slices (i.e. regions of the sample) which are highlighted visually. Note that logically this still does nothing, and is the major focus going forward. The tagging system and waveform system are not integrated together at all, minus the visual component of it.