Add tag & waveform integration
Once the saving features are added, this will be the final version for the demo. The tagging and waveform have been integrated only for the hardcoded sample to be used for the demo, however it is very helpful insight into how to integrate the two generally after the demo. Additionally, a separate class containing marker points and slice descriptions has been added to emulate the features when it has all been finished. If you use the sample "DL Broken.wav" that we are using for the demo, it will function exactly as it should minus the saving and some future ease-of-use features that have fallen to the backburner.